P1120R0 - part 5

+++ imgui.cpp
@@ -2298,7 +2298,7 @@ int ImTextCharFromUtf8(unsigned int* out_char, const char* in_text, const char* 
         // One byte is consumed in case of invalid first byte of in_text.
         // All available bytes (at most `len` bytes) are consumed on incomplete/invalid second to last bytes.
         // Invalid or incomplete input may consume less bytes than wanted, therefore every byte has to be inspected in s.
-        wanted = ImMin(wanted, !!s[0] + !!s[1] + !!s[2] + !!s[3]);
+        wanted = ImMin(wanted, (int)(!!s[0]) + (int)(!!s[1]) + (int)(!!s[2]) + (int)(!!s[3]));
         *out_char = IM_UNICODE_CODEPOINT_INVALID;
     }
